recent: Rename recent_topics_util.
