+++ web/src/compose_actions.js
@@ -22,7 +22,7 @@ import * as narrow_state from "./narrow_state";
 import {page_params} from "./page_params";
 import * as people from "./people";
 import * as recent_view_ui from "./recent_view_ui";
-import * as recent_topics_util from "./recent_topics_util";
+import * as recent_view_util from "./recent_view_util";
 import * as reload_state from "./reload_state";
 import * as resize from "./resize";
 import * as settings_config from "./settings_config";
@@ -307,7 +307,7 @@ export function cancel() {
 export function respond_to_message(opts) {
     let message;
     let msg_type;
-    if (recent_topics_util.is_visible()) {
+    if (recent_view_util.is_visible()) {
         message = recent_view_ui.get_focused_row_message();
         if (message === undefined) {
             // Open empty compose with nothing pre-filled since
