+++ web/src/message_lists.js
@@ -2,7 +2,7 @@ import $ from "jquery";
 
 import {Filter} from "./filter";
 import * as message_list from "./message_list";
-import * as recent_topics_util from "./recent_topics_util";
+import * as recent_view_util from "./recent_view_util";
 import * as ui_util from "./ui_util";
 
 export let home;
@@ -14,7 +14,7 @@ export function set_current(msg_list) {
 
 export function all_rendered_message_lists() {
     const rendered_message_lists = [home];
-    if (current !== home && !recent_topics_util.is_visible()) {
+    if (current !== home && !recent_view_util.is_visible()) {
         rendered_message_lists.push(current);
     }
     return rendered_message_lists;
