+++ web/src/recent_view_ui.js
@@ -30,7 +30,7 @@ import * as pm_list from "./pm_list";
 import * as popovers from "./popovers";
 import * as recent_senders from "./recent_senders";
 import {get, process_message, topics} from "./recent_topics_data";
-import {get_key_from_message, get_topic_key, is_visible, set_visible} from "./recent_topics_util";
+import {get_key_from_message, get_topic_key, is_visible, set_visible} from "./recent_view_util";
 import * as resize from "./resize";
 import * as scroll_util from "./scroll_util";
 import * as search from "./search";
