+++ web/src/recent_view_util.js
