+++ web/tests/compose_actions.test.js
@@ -42,7 +42,7 @@ const narrow_state = mock_esm("../src/narrow_state", {
 mock_esm("../src/reload_state", {
     is_in_progress: () => false,
 });
-mock_esm("../src/recent_topics_util", {
+mock_esm("../src/recent_view_util", {
     is_visible: noop,
 });
 mock_esm("../src/drafts", {
