+++ web/tests/compose_closed_ui.test.js
@@ -7,7 +7,7 @@ const {mock_esm, set_global, zrequire} = require("./lib/namespace");
 const {run_test} = require("./lib/test");
 const $ = require("./lib/zjquery");
 
-mock_esm("../src/recent_topics_util", {
+mock_esm("../src/recent_view_util", {
     is_visible: () => false,
 });
 const noop = () => {};
